improve verdy publish steps

+++ packages/boarding/src/verdy/index.ts
@@ -1,7 +1,9 @@
 import { execAsync } from '@goatjs/node/exec';
-import { checkGitStatus } from './helpers.js';
+import { checkGitStatus, isMonorepo } from './helpers.js';
 import { git } from '../git/index.js';
 import { rimraf } from '@goatjs/rimraf';
+import { publish, type PublishOptions } from './publish.js';
+import { consoleColor } from '@goatjs/node/console-color';
 
 // TODO [2025-07-30] IMPORTANT breaking change,for monorepo use monorepo.publish and create a cli where you can do:
 // pass a scope like @goatjs. and do yarn verdaccio publish pslist boarding core node ecc or yarn verdaccio publish all to publish all.
@@ -9,33 +11,21 @@ import { rimraf } from '@goatjs/rimraf';
 
 // TODO [2025-07-30] use spawn to run publish and add more consoleColors
 
-interface PublishOptions {
-  version?: 'major' | 'minor' | 'patch';
-}
-
 export const verdy = {
-  publish: async ({ version = 'patch' }: PublishOptions = {}) => {
+  publish: async ({ version }: PublishOptions = {}) => {
+    consoleColor('yellow', "Verdy detect that you're running in a monorepo. Please ensure to run this scripts from the root only.");
     await checkGitStatus();
-    await execAsync('yarn build'); // test before increasing version.
-    await execAsync(`yarn version ${version}`);
-    await execAsync('yarn npm publish');
+    await publish({ version, isMonorepo: await isMonorepo() });
     await git.add();
     await git.commit('RELEASE');
     await git.push();
   },
-  monorepo: {
-    publish: async (_packages?: string[], { version = 'patch' }: PublishOptions = {}) => {
-      await checkGitStatus();
-      await execAsync('yarn build'); // test before increasing version.
-      await execAsync(`yarn workspaces foreach --all --no-private version ${version}`);
-      await execAsync('yarn workspaces foreach --all --no-private npm publish');
-      await git.add();
-      await git.commit('RELEASE');
-      await git.push();
-    },
-    clear: async () => {
+  clear: async () => {
+    if (await isMonorepo()) {
       await rimraf('.turbo');
       await execAsync('yarn workspaces foreach --all run rimraf dist .turbo .next');
-    },
+    } else {
+      await execAsync('yarn rimraf dist .next');
+    }
   },
 };
